pyproject.toml
[tool.poetry]
name = "ollama_rag"
version = "2.2.6"
description = ""
authors = ["Marc Mezger <marc.mezger@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
langchain = "^0.0.339"
streamlit = "^1.28.2"
loguru = "^0.7.2"
omegaconf = "^2.3.0"
pypdf = "^3.17.0"
python-dotenv = "^1.0.0"
python-multipart = "^0.0.6"
grpcio = "^1.59.3"
grpcio-tools = "^1.59.3" # for qdrant
qdrant-client = "^1.6.9" # for qdrant
nltk = "^3.8.1"
pypdfium2 = "^4.24.0"
chainlit = "^0.7.700"
langchainhub = "^0.1.14"


[tool.poetry.group.dev.dependencies]
mypy = "^1.4.1"
pytest = "^7.4.0"
pylint = "^2.17.4"
ruff = "^0.0.280"
pytest-asyncio = "^0.21.0"
coverage = "^7.2.7"
types-urllib3 = "^1.26.25.14"
lxml-stubs = "^0.4.0"
pandas-stubs = "^2.0.2.230605"
types-pillow = "^10.0.0.2"
types-pymysql = "^1.1.0.1"
types-pyyaml = "^6.0.12.11"
types-beautifulsoup4 = "^4.12.0.5"
types-cachetools = "^5.3.0.6"
types-google-cloud-ndb = "^2.1.0.8"
types-protobuf = "^4.23.0.2"
types-psutil = "^5.9.5.16"
types-psycopg2 = "^2.9.21.11"
types-python-dateutil = "^2.8.19.14"
types-pytz = "^2023.3.0.0"
types-redis = "^4.6.0.3"
types-regex = "^2023.6.3.1"
types-requests = "^2.31.0.2"
types-toml = "^0.10.8.7"
types-tqdm = "^4.65.0.2"
types-tzlocal = "^5.0.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.settings]
known_third_party = ["aleph_alpha_client", "dotenv", "fastapi", "gpt4all", "httpx", "langchain", "loguru", "nltk", "numpy", "omegaconf", "openai", "pydantic", "pytest", "qdrant_client", "requests", "starlette", "streamlit", "tqdm"]

[tool.mypy]
# Disallow dynamic typing
disallow_any_unimported=true
disallow_any_expr = true
disallow_any_decorated = true
disallow_any_generics = true
disallow_any_explicit = true
disallow_subclassing_any = true

# Disallow untyped definitions and calls
disallow_untyped_calls = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true

# None and optional handling
no_implicit_optional = true

# Configuring warnings
warn_unused_ignores = true
warn_no_return = true
warn_return_any = true
warn_redundant_casts = true

# Misc things
strict_equality = true

# Config file
warn_unused_configs = true

[tool.ruff]
line-length = 170
[tool.pylint]
max-line-length = 170
